Test: rearrange tests
