+++ test/fixtures/array.js
